Fixed finished generation event
The event for whenever the generation of the octree is finished now gets thrown correctly.
Cleaned up code.
Moved structs and enums to seperate files to increase file structure.
Nodes now know to which octree they belong.
Added seperate layer for the GeometryNodes to allow them to only hit objects on the same layer and the Geometry layer. ( Too reduce possible collision bugs)