Fixed Inconsistend Finished Generation Events
Finally, after many thinking to have fixed the problem many times, I have revamped the generation system to no longer use FixedUpdat() and OnTriggerEnter(). It now detects colliders in its Start() and does all necessary checks immediatly. This solves all other problems related to this detection aswel.
Also simplified the code quite a bit due to this change.
Added a collider and rigidbody cleanup to the generator to smoothen the in-editor experience.