Added correct saving of the exits & introduced MaxDiameter
The exits get correctly saved in the appropriate tunnel or hole.
The sepperate lists for tunnels and holes have been merged into 1 list.
The detection process has been finetuned. Now, everything is a tunnel. Later on, if a hole has more than 1 exit, it is changed to a tunnel.
Introduced a variable to limit the maximum diameter a tunnel or hole may be.